% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_niche_factor.R
\name{as_niche_factor}
\alias{as_niche_factor}
\title{Converts a vector or matrix of quantitative environmental data as a 
factorized version of a desired resolution}
\usage{
as_niche_factor(env, resolution = 3)
}
\arguments{
\item{env}{A matrix or data.frame of quantitative environmental data}

\item{resolution}{Number of intervals in which divide the range of each variable}
}
\value{

}
\description{
Converts a vector or matrix of quantitative environmental data as a 
factorized version of a desired resolution
}
\examples{

}
