% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_species.R
\name{espindic}
\alias{espindic}
\title{Performs a species indicator analysis following  
Dufrene, M. and P. Legendre. 1997. Species assemblages and indicator species:
The need for a flexible asymmetrical approach. Ecological Monographs 67: 345-366 
McCune, B. & Grace, J. B. Analysis of Ecological Communities 2002, 300 (page 198)}
\usage{
espindic(com, group)
}
\arguments{
\item{com}{A matrix of community data}

\item{group}{A vector indicating grouping of observations in the community data}
}
\value{

}
\description{
Performs a species indicator analysis following  
Dufrene, M. and P. Legendre. 1997. Species assemblages and indicator species:
The need for a flexible asymmetrical approach. Ecological Monographs 67: 345-366 
McCune, B. & Grace, J. B. Analysis of Ecological Communities 2002, 300 (page 198)
}
