% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_species.R
\name{indicator_species}
\alias{indicator_species}
\title{Performs a species indicator analysis}
\usage{
indicator_species(com, group, times = 100)
}
\arguments{
\item{com}{A matrix of community data}

\item{group}{A vector indicating grouping of observations in the community data}
}
\value{
The indicator species
}
\description{
@references 
  
  Dufrene, M. and P. Legendre. 1997. Species assemblages and indicator species:
The need for a flexible asymmetrical approach. Ecological Monographs 67: 345-366
}
\details{
McCune, B. & Grace, 
J. B. Analysis of Ecological Communities 2002, 300 (page 198)
}
\examples{

data(soilandfauna)

# Select community (species) data

com <- soilandfauna[,18:50]

# Select Grouping factor

group <- soilandfauna[,1]

out <- indicator_species(com,group)

barplot(out)

}
