% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_indicator_species.R
\name{select_indicator_species}
\alias{select_indicator_species}
\title{Select indicator species}
\usage{
select_indicator_species(com, group, alfa = 0.05)
}
\arguments{
\item{com}{Community data. matrix or data.frame with species in columns and samples in rows.}

\item{group}{Vector of the sample grouping of the community data.}

\item{alfa}{Significance level used for the test.}
}
\value{
A list with two components: a vector of the name of the 
indicator species selected 
and a vector of the conditional probability of each species 
or taxonomic unit in the data.
}
\description{
Selection of species or taxonomic units, depending on the taxonomic resolution
 in the original data set, that indicate with a given probability of
belonging to an environment.
}
\examples{

data(soilandfauna)
 
# Select community (species) data
com <- soilandfauna[,18:50]

# Select grouping factor
group <- soilandfauna[,1]

select_indicator_species(com,group,0.05)

}
