% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_env.R
\name{identify_env}
\alias{identify_env}
\title{Identify the environment from new samples of community species}
\usage{
identify_env(com, com.to.identify, group, alfa = 0.05)
}
\arguments{
\item{com}{The original community used to select indicator species.}

\item{com.to.identify}{A new sample o group of samples to identify the 
environment that they belong.}

\item{group}{vector of the samples grouping}

\item{alfa}{Significance level used for the test.}
}
\value{
A list with the the environemnt estimation value for each group 
and the belonging environment estimated.
}
\description{
This function use the result of `select_indicator_species` to identify a new 
set of samples of the community
}
\examples{

data(soilandfauna)
 
# Select community (species) data

com <- soilandfauna[,18:50]

# Read data of new community matrix with abundances of the same
# species as 'com'.

newcom <- read.delim("data/newcom.txt")

# Select grouping factor
group <- soilandfauna[,1]

identify_env(com, newsamples, group)

}
