% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_indicator_species.R
\name{select_indicator_species}
\alias{select_indicator_species}
\title{Select indicator species}
\usage{
select_indicator_species(com, group, alfa = 0.05)
}
\arguments{
\item{com}{matrix or data.frame with species in columns and samples in rows}

\item{group}{vector of the environments of the samples}

\item{alfa}{significance level used for the test}
}
\value{
A data.frame with the species that are above
}
\description{
This function select species that indicate with a given probability the 
pertenence to an environment using a matrix of species samples at different
environments using a chisqrt test with a given significance level.
}
\examples{
da <- read.delim("Data/data.txt")
############################# CHANGE THIS TO EMBEDDED PACKAGE DATA

# Select community (species) data
com <- da[,18:50]
# Select environmental data
env <- da[,3:17]

# Grouping factor
group <- da[,1]

select_indicator_species(com,group,0.05)

}
