% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_hypercube.R
\name{sp_hypercube}
\alias{sp_hypercube}
\title{A partir de un vector o matriz de datos ambientales
devuelve la presencia de una especie en hipercubos de una resolución definida}
\usage{
sp_hypercube(env, com, resolution = 3)
}
\arguments{
\item{env}{Vector or matrix of quantitative environmental data}

\item{com}{A vector of abundance of a species registered in the same 
samples as environmetal data}

\item{resolution}{To divide the range of each environmental variable}
}
\value{

}
\description{
A partir de un vector o matriz de datos ambientales
devuelve la presencia de una especie en hipercubos de una resolución definida
}
\examples{


}
